+++ adam/entities/__init__.py
@@ -1,3 +1,3 @@
 from ..core.entities import Configuration
-from ..features.simulation.entities import Data, Collision, CollisionData, ConfigurationData
+from ..features.simulation.entities import AdamInfo, Collision, ManipulatorData
 from ..features.map_maker.entities import Body, Box, Capsule, Cube, Cylinder, Sphere