+++ adam/features/simulation/use_cases/data_manager.py
@@ -1,30 +1,34 @@
 import numpy as np
 
 from .collision_detector import CollisionDetector
-from ..entities import Data, CollisionData, ConfigurationData, Collision, Configuration
+from ..entities import AdamInfo, ManipulatorData, Collision, Configuration, System
 
 
 class DataManager:
     def __init__(self) -> None:
-        self.configurations: np.ndarray = None  # type:ignore
-        self.geometry_1_list: np.ndarray = None  # type:ignore
-        self.geometry_2_list: np.ndarray = None  # type:ignore
-
         self.collision_detector: CollisionDetector = CollisionDetector()
 
-    def update(self, data) -> None:
-        self.configurations = data.qpos.copy()
-        self.geometry_1_list = data.contact.geom1.copy()
-        self.geometry_2_list = data.contact.geom2.copy()
+    def get(self, data) -> AdamInfo:
+        # Configurations
+        left_configuration: Configuration = Configuration(*data.qpos[0:6])
+        right_configuration: Configuration = Configuration(*data.qpos[6:12])
 
-    def get(self) -> Data:
-        left_manipulator_configuration: Configuration = Configuration(*self.configurations[0:6])
-        right_manipulator_configuration: Configuration = Configuration(*self.configurations[6:12])
+        # Collisions
+        left_collision: Collision = self.collision_detector.check_left_manipulator(data.contact.geom1, data.contact.geom2)
+        right_collision: Collision = self.collision_detector.check_right_manipulator(data.contact.geom1, data.contact.geom2)
 
-        left_manipulator_collision: Collision = self.collision_detector.check_left_manipulator(self.geometry_1_list, self.geometry_2_list)
-        right_manipulator_collision: Collision = self.collision_detector.check_right_manipulator(self.geometry_1_list, self.geometry_2_list)
+        # Systems
+        left_systems: list[System] = [System.from_htm(np.array([[*orientation[0:3], position[0]],
+                                                                [*orientation[3:6], position[1]],
+                                                                [*orientation[6:9], position[2]]]))
+                                      for orientation, position in zip(data.xmat[2:9], data.xpos[2:9])]
 
-        configuration_data: ConfigurationData = ConfigurationData(left_manipulator_configuration, right_manipulator_configuration)
-        collision_data: CollisionData = CollisionData(left_manipulator_collision, right_manipulator_collision)
+        right_systems: list[System] = [System.from_htm(np.array([[*orientation[0:3], position[0]],
+                                                                [*orientation[3:6], position[1]],
+                                                                 [*orientation[6:9], position[2]]]))
+                                       for orientation, position in zip(data.xmat[9:], data.xpos[9:])]
 
-        return Data(configuration_data, collision_data)
+        left_manipulator_data: ManipulatorData = ManipulatorData(left_collision, left_configuration, left_systems)
+        right_manipulator_data: ManipulatorData = ManipulatorData(right_collision, right_configuration, right_systems)
+
+        return AdamInfo(left_manipulator_data, right_manipulator_data)