+++ tests/test_2.py
@@ -1,17 +1,17 @@
 from adam import Simulation, ConfigurationsManager
-from adam.entities import Configuration, Data
+from adam.entities import Configuration, AdamInfo
 
 
 def main():
     sim: Simulation = Simulation()
-    initial_data: Data = sim.load_scene()
+    initial_info: AdamInfo = sim.load_scene()
 
     configuration_list: list[Configuration] = ConfigurationsManager.load('test')
 
     for configuration in configuration_list:
         sim.render()
 
-        data: Data = sim.step(configuration, initial_data.configuration.right_manipulator)
+        info: AdamInfo = sim.step(configuration, initial_info.right_manipulator.configuration)
 
     sim.close()